+++ optaplanner-examples/src/main/resources/org/optaplanner/examples/nqueens/optional/benchmark/nqueensBenchmarkConfig.xml
