+++ optaplanner-examples/src/main/resources/org/optaplanner/examples/nqueens/solver/nQueensConstraints.drl
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.nqueens.solver;
+package org.optaplanner.examples.nqueens.optional.solver;
     dialect "java"
 
 import org.optaplanner.core.api.score.buildin.simple.SimpleScoreHolder;
