+++ optaplanner-examples/src/test/java/org/optaplanner/examples/nqueens/optional/solver/score/NQueensConstraintProviderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.examples.nqueens.solver.score;
+package org.optaplanner.examples.nqueens.optional.solver.score;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -25,6 +25,7 @@ import org.optaplanner.examples.nqueens.domain.Column;
 import org.optaplanner.examples.nqueens.domain.NQueens;
 import org.optaplanner.examples.nqueens.domain.Queen;
 import org.optaplanner.examples.nqueens.domain.Row;
+import org.optaplanner.examples.nqueens.optional.score.NQueensConstraintProvider;
 import org.optaplanner.persistence.xstream.impl.domain.solution.XStreamSolutionFileIO;
 import org.optaplanner.test.api.score.stream.ConstraintVerifier;
