+++ optaplanner-examples/src/test/resources/org/optaplanner/examples/nqueens/optional/solver/score/256queensScore-30.xml
